Move 'FeedStore' to it's own file
