+++ EssentialFeed/EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -38,12 +38,6 @@ public class LocalFeedLoader {
     }
 }
 
-public protocol FeedStore {
-    typealias DeletionCompletion = (Error?) -> Void
-    typealias InsertionCompletion = (Error?) -> Void
 
-    func deleteCachedFeed(completion: @escaping DeletionCompletion)
-    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
-}
